Fix error handling for already created functions that could give syntax error in newer releases (#3250)
Our code assumes that function once created will never give syntax error. But certain fixes like 5251e4a where we disallow declaring reserved @@var, functions could have been created in earlier releases which will give syntax error while executing on a version with this commit. While throwing this syntax error at time of execution of function we have not cleaned up the contexts and aborted well which leads to a crash while executing a subsequent query.
To fix this we call terminate_batch for compile errors when we get parser syntax error.

In this PR We were crashing due to this in this action and by adding this commit we are not crashing anymore.

Issues Resolved
BABEL-5455