Optimize DDL export framework execution time (#3266)
Previously, the DDL export framework was taking over 20 minutes to run the ddl_all_objects.sql test, while other tests completed within a minute. Upon further investigation, we found that we were not filtering out objects in the 'sys' schema before checking IsSystemObject. This resulted in executing a query for each object present in the 'sys' schema to determine whether it was a system object or not.

This commit adds a schema filter before the IsSystemObject check in the SMO script. By first excluding objects in the 'sys' schema, we avoid unnecessary IsSystemObject checks on system objects, significantly reducing execution time of framework.

Task: BABEL-5473

Signed-off-by: Sumit Jaiswal sumiji@amazon.com