Fix alter table drop column drops trigger on the table (#3225) (#3276)
In postgres trigger and functions are two different objects. You first define a function and then use it in your trigger definition. In babelfish we handle this internally so that for users it feels like a single object. As part of this internal handing we also drop the function created for this trigger as part of post object access hook. Currently the condition for DROP TABLE was incomplete and even for DROP COLUMN we would drop the triggers associated with the table.
As a fix add another condition to only execute this internal drop of triggers for DROP TABLE.

Issues Resolved: BABEL-5417
Signed-off-by: Tanzeel Khan <tzlkhan@amazon.com>