Fix the incorrect string limit when declaring nvarchar(max) (#3271)
Previously, the limit on the max length of NVARCHAR(Max) was 4000, longer characters will be truncated. But NVARCHAR(MAX) should be capable of handling strings far longer than 4,000 characters. This pr change the limit to default INT_MAX rather than 4000 to fix this issue, and add test cases to verify.

The issue is the typemod for nvarchar(max) in AS JSON queries is set as 4000, which limits the query length, the solution is to remove the code setting 4000 and recover typemod from TSQLMaxTypmod to -1 in tds client.

Task:BABEL-5054
Signed-off-by: Chenxiao Wang <chxwang@amazon.com>